20230617增加复式，胆拖前端接口

+++ zlt-business/marksix-center/marksix-server/src/main/java/com/central/marksix/entity/dto/QuizSubordersDto.java
@@ -26,5 +26,5 @@ public class QuizSubordersDto {
     @ApiModelProperty(value = "注数")
     private Integer units;
     @ApiModelProperty(value = "赔率")
-    private BigDecimal odds;
+    private Integer odds;
     }
